refactor(Player): 로그인 관련 부분 분리
